feat(branches): release from the branch "main" as well as "master"
We are working towards removing use of terms like "master". To this end we will be renaming our
"master" branches to "main", and releases will be performed from the "main" branch. For backwards
compatibility we will still release from "master" for repositories that have not yet been updated.
Once the transition is complete we will only perform releases from the "main" branch.